backend/requirements.txt
django==3.0.5
djangorestframework==3.11.0
psycopg2==2.8.5
django-rest-auth[with_social]==0.9.5
drf-extensions==0.6.0
